requirements.txt
pyyaml
openai
tqdm
tenacity
python-dotenv
pandas
openpyxl
tiktoken
pdfplumber